PR(REFACTOR): Split large p2p file into smaller ones
